medication grade front edit

+++ backend/src/main/java/app/service/impl/PharmacyServiceImpl.java
@@ -87,7 +87,7 @@ public class PharmacyServiceImpl implements PharmacyService {
                         pmDTO.setAddress(p.getAddress());
                         pmDTO.setMedicationId(medicationId);
 
-                        double cena = medicationPriceListService.getMedicationPrice(p.getId(), medicationId);
+                        double cena = (Double)medicationPriceListService.getMedicationPrice(p.getId(), medicationId);
                         pmDTO.setMedicationPrice(cena);
                         pharmacies.add(pmDTO);
                     }
